untangle some guide components
