+++ __tests__/components/FeedbackOverview.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render, screen } from "@testing-library/react";
 import { useGuide } from "providers/GuideProvider";
-import { FeedbackOverview } from "../../app/components/guideCard/FeedbackOverview";
+import { FeedbackOverview } from "../../app/components/feedbackOverview/FeedbackOverview";
 import { FeedbackDocumentWithReturn } from "types/guideTypes";
 import { ReturnDocument } from "models/return";
 import { FeedbackDocument } from "models/review";
