+++ app/components/GiveFeedbackView.tsx/style.tsx
@@ -1,0 +1,6 @@
+import { FlexibleWrapper } from "globalStyles/globalStyles";
+import styled from "styled-components";
+
+export const FeedbackInfoContainer = styled(FlexibleWrapper)`
+  gap: 2rem;
+`;
