+++ app/components/guidesClient/style.tsx
@@ -1,0 +1,10 @@
+"use client";
+import styled from "styled-components";
+
+export const GuideContainer = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-around;
+  gap: 0.75rem;
+  width: 100%;
+`;
