+++ app/globalStyles/globalConstants.ts
@@ -1,0 +1,1 @@
+export const BREAKPOINT_DESKTOP = "1024px";
